// cinnamon-platform/cinnamon-frontend/src/app/shared/model/synthetization-component-progress.ts
export class SynthetizationComponentProgress {
    completed: string;
    duration: string | null;
    fitting_duration: string | null;
    fitting_remaining_time?: string | null;
    initialization_duration: string | null;
    remaining_time: string | null;
    sampling_duration: string | null;
    sampling_remaining_time?: string | null;
    synthesizer_name: string | null;
}